One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/planning/rsstorelist/rsstorelist.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { RsstorelistComponent } from './rsstorelist.component';

describe('Component: Rsstorelist', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [RsstorelistComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([RsstorelistComponent],
      (component: RsstorelistComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(RsstorelistComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(RsstorelistComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-rsstorelist></app-rsstorelist>
  `,
  directives: [RsstorelistComponent]
})
class RsstorelistComponentTestController {
}

